Move internal class from Polyhedron to BGL
