+++ Polyhedron/include/CGAL/boost/graph/properties_Polyhedron_3.h
@@ -103,42 +103,6 @@ struct Wrap_squared
   }
 };
 
-  template<typename Polyhedron, typename Handle>
-struct Index_accessor
-    : boost::put_get_helper< std::size_t&, Index_accessor<Polyhedron,Handle> >
-{
-  typedef boost::lvalue_property_map_tag category;
-  typedef std::size_t&                   reference;
-  typedef std::size_t                    value_type;
-  typedef Handle                         key_type;
-
-  reference operator[](Handle h) const { return h->id(); }
-};
-
-template<typename Handle>
-struct Edge_index_accessor
-  : boost::put_get_helper< std::size_t, Edge_index_accessor<Handle> >
-{
-  typedef boost::readable_property_map_tag category;
-  typedef std::size_t                      reference;
-  typedef std::size_t                      value_type;
-  typedef Handle                           key_type;
-
-  reference operator[](Handle h) const { return h.id(); }
-};
-
-template<typename Handle, typename ValueType, typename Reference>
-struct Point_accessor
-  : boost::put_get_helper< Reference, Point_accessor<Handle, ValueType, Reference> >
-{
-  typedef boost::lvalue_property_map_tag category;
-  typedef Reference                      reference;
-  typedef ValueType                      value_type;
-  typedef Handle                         key_type;
-
-  reference operator[](Handle h) const { return h->point(); }
-};
-
 } // internal
 
 // the tag we dispatch on from property_map<G, Property>
